refactor(compaction): isolate execution machinery
